requirements.txt
google-cloud-bigquery>=3.11.0
google-cloud-bigquery[pandas]>=3.11.0
pandas>=2.0.0
datasets>=2.18.0,<4.0.0
huggingface-hub>=0.25.0
python-dotenv>=1.0.0
pyarrow>=15.0.0
db-dtypes>=1.1.0
tqdm>=4.66.0
matplotlib>=3.7.0
transformers>=4.38.0
torch>=2.0.0
google-generativeai>=0.5.0
sentencepiece>=0.1.99
accelerate>=0.27.0
